Bugfix: flattening for non-binary bitnor, bitnand, bitxnor
This fixes the flattening for bitnor, bitnand, bitxnor for the case that the
expression has one operand or more than two operands.

For one operand, the result is now the bit-wise negation.

For three or more operands, the result is now the bit-wise negation of
the bitor, bitand, bitxor with the same operands.